Revert "Refactor canvas viewer"
