+++ package.json
@@ -27,7 +27,6 @@
     "@emailjs/browser": "^4.4.1",
     "@hello-pangea/dnd": "^18.0.1",
     "@hookform/resolvers": "^4.1.3",
-    "@hyzyla/pdfium": "^2.1.9",
     "@iiif/helpers": "^1.3.2",
     "@iiif/parser": "^2.2.0",
     "@iiif/presentation-3": "^2.2.3",
@@ -48,7 +47,7 @@
     "@radix-ui/react-radio-group": "^1.3.8",
     "@radix-ui/react-scroll-area": "^1.2.10",
     "@radix-ui/react-select": "^2.1.6",
-    "@radix-ui/react-separator": "^1.1.8",
+    "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.2.0",
     "@radix-ui/react-switch": "^1.1.3",
     "@radix-ui/react-tabs": "^1.1.3",
@@ -117,11 +116,10 @@
     "sonner": "^2.0.1",
     "tailwind-merge": "^3.0.1",
     "tailwindcss-animate": "^1.0.7",
-    "tesseract.js": "^6.0.1",
+    "tesseract.js": "^6.0.0",
     "uuid": "^11.1.0",
     "vaul": "^1.1.2",
-    "xlsx": "^0.18.5",
-    "zustand": "^5.0.8"
+    "xlsx": "^0.18.5"
   },
   "devDependencies": {
     "@babel/preset-typescript": "^7.26.0",
@@ -152,7 +150,6 @@
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.18",
-    "file-system-access": "^1.0.4",
     "globals": "^15.15.0",
     "jsdom": "^26.0.0",
     "postcss": "^8.5.2",
