+++ public/locales/fr/translation.json
@@ -86,7 +86,6 @@
   "btn_login": "Se connecter",
   "btn_reset_indexeddb": "Réinitialiser IndexedDB",
   "btn_experimental_features": "Fonctionnalités expérimentales",
-  "btn_select_folder": "Sélectionner un dossier",
   "link_open_element": "ouvrir l'élément",
   "form_placeholder_key": "Une clé",
   "form_placeholder_value": "Une valeur",
@@ -163,8 +162,6 @@
   "title_remove_annotations": "Supprimer des annotations",
   "title_are_you_sure": "Êtes-vous sûr ?",
   "title_export_worker_result": "Exporter le résultat de {{name}}",
-  "title_use_local_manifest": "Manifest local",
-  "title_pdf_in_folder": "Fichiers pdf dans le dossier sélectionné :",
   "table_col_title_key": "Clé",
   "table_col_title_value": "Valeur",
   "table_col_title_tags": "Tags",
@@ -217,7 +214,6 @@
   "error_unsupported_file_type": "Type de fichier non supporté. Veuillez utiliser un fichier de type : {{types}}.",
   "error_mode_name_exists": "Un modèle avec le nom {{name}} existe déjà.",
   "error_model_undefined": "Aucun modèle de données n'est associé à cette collection. Veuillez en sélectionner un dans le formulaire de détails de la collection.",
-  "error_fs_access_not_supported": "L'API File System Access n'est pas supportée par votre navigateur. Essayez avec un autre navigateur (Chrome ou Edge).",
   "info_empty_collection": "Cette collection est vide",
   "info_empty_manifest": "Ce manifest ne contient aucune page",
   "info_no_collection": "Vous n'avez aucune collection actuellement",
@@ -257,7 +253,6 @@
   "info_selection_count": "{{count}} élément(s) sélectionné(s)",
   "info_experimental_features": "Fonctionnalités expérimentales. Ces fonctionnalités sont en cours de développement actif et peuvent être instables.",
   "info_reset_indexeddb": "La réinitialisation d'IndexedDB est irréversible et entrainera la perte de toutes les données Corpusense stockées sur ce navigateur (collections, modèles, etc.).",
-  "info_empty_folder": "Le dossier sélectionné est vide.",
   "description_select_model": "Choisissez un modèle de données",
   "description_import_collection": "Importer une collection (fichier JSON) ou un ensemble de collections (fichier ZIP)",
   "description_create_collection": "Créer une collection en saisissant le nom de la collection.",
@@ -269,7 +264,6 @@
   "description_delete_model": "Êtes-vous sûr de vouloir supprimer ce modèle ? Cette action est irréversible",
   "description_delete_datafield": "Êtes-vous sûr de vouloir supprimer ce champ ? Cette action est irréversible",
   "description_select_export_formats": "Sélectionnez les formats d'export souhaités :",
-  "description_local_manifest": "Sélectionnez un dossier contenant soit un pdf, soit une ou plusieurs images (jpg, png, tiff). Corpusense générera un manifest IIIF à partir de ces fichiers.",
   "menu_create_from_selection": "Créer une collection à partir de la sélection",
   "menu_add_selection_to_collection": "Ajouter la sélection à une collection",
   "menu_define_start": "Définir le début de sélection ici",
