+++ src/App.tsx
@@ -2,6 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { AlertDialogProvider } from './components/reducers/AlertDialogContext';
 import { CorpusenseRoutes } from './hooks/useAppNavigation';
 import { ExperimentalProvider } from './hooks/useExperimental';
+import useServiceWorker from './hooks/useServiceWorker';
 import { initI18n } from './i18n';
 import CollectionInspectorPage from './pages/CollectionInspectorPage';
 import CollectionsManagerPage from './pages/CollectionsManagerPage';
@@ -29,11 +30,10 @@ initI18n()
   });
 
 function App() {
-  // useServiceWorker();// Hook to manage service worker updates (cached files updates)
+  useServiceWorker();
 
   return (
     <BrowserRouter basename={basePath}>
-      {/* <LocalManifestProvider> */}
       <ExperimentalProvider>
         <AlertDialogProvider>
           <Routes>
@@ -62,7 +62,6 @@ function App() {
           </Routes>
         </AlertDialogProvider>
       </ExperimentalProvider>
-      {/* </LocalManifestProvider> */}
     </BrowserRouter>
   );
 }
