+++ src/components/QuickCanvasViewer.tsx
@@ -16,7 +16,7 @@ import { useAppSelector } from '@/hooks/hooks';
 import { selectCollections } from '@/state/selectors/collections';
 import { Canvas } from '@iiif/presentation-3';
 import { useEffect, useState } from 'react';
-import CanvasViewer from './canvasViewer/CanvasViewer';
+import CanvasViewer from './CanvasViewer';
 
 const QuickCanvasViewer = () => {
   const collections = useAppSelector(selectCollections);
