--- src/data/models/FSHandle.ts
@@ -1,4 +1,0 @@
-export interface FSHandle {
-  id: string;
-  handle: FileSystemDirectoryHandle;
-}
