+++ src/state/sagas/plugins/workers/peroocr.ts
@@ -116,7 +116,7 @@ export default async function run(task: Task, _params: PluginParams): Promise<Wo
 
 export function exportResult(results: Result[], formats: string[]) {
   if (results.length === 0) {
-    console.warn(`No results to export from ${pluginDisplayName} plugin`);
+    console.warn('No results to export from Mistral plugin');
     return;
   }
