+++ src/utils/manifest.ts
@@ -29,30 +29,22 @@ export type CanvasInfo = {
   height: number;
 };
 
-export const generateManifest = ({
-  documentName,
-  canvasInfo,
-  folder,
-  manifestId,
-  isFileSystem = false,
-}: {
-  documentName: string;
-  canvasInfo: CanvasInfo[];
-  folder: string;
-  manifestId?: string;
-  isFileSystem?: boolean;
-}): Manifest => {
-  const url_from = isFileSystem ? '' : `${import.meta.env.VITE_SUPABASE_STORAGE_URL}/${folder}/`;
+export const generateManifest = (
+  documentName: string,
+  canvasInfo: CanvasInfo[],
+  folder: string,
+): Manifest => {
+  const url_supabase = `${import.meta.env.VITE_SUPABASE_STORAGE_URL}/${folder}/`;
 
   return {
     '@context': 'http://iiif.io/api/presentation/3/context.json',
-    id: manifestId ?? `${url_from}/manifest.json`,
+    id: `${url_supabase}/manifest.json`,
     type: 'Manifest',
     label: {
       fr: [documentName],
     },
     items: canvasInfo.map((canvas, index) => ({
-      id: `${url_from}/canvas/p${index + 1}`,
+      id: `${url_supabase}/canvas/p${index + 1}`,
       type: 'Canvas',
       label: {
         fr: [`Page ${index + 1}`],
@@ -68,11 +60,11 @@ export const generateManifest = ({
       ],
       items: [
         {
-          id: `${url_from}/page/p${index + 1}/1`,
+          id: `${url_supabase}/page/p${index + 1}/1`,
           type: 'AnnotationPage',
           items: [
             {
-              id: `${url_from}/annotation/p${index + 1}/1-image`,
+              id: `${url_supabase}/annotation/p${index + 1}/1-image`,
               type: 'Annotation',
               motivation: 'painting',
               body: {
@@ -81,15 +73,13 @@ export const generateManifest = ({
                 format: 'image/png',
                 width: Math.floor(canvas.width),
                 height: Math.floor(canvas.height),
-                service: isFileSystem
-                  ? []
-                  : [
-                      {
-                        id: canvas.id.substring(0, canvas.id.length - 23), // Retirer l'extension .png
-                        type: 'ImageService3',
-                        profile: 'level1',
-                      },
-                    ],
+                service: [
+                  {
+                    id: canvas.id.substring(0, canvas.id.length - 23), // Retirer l'extension .png
+                    type: 'ImageService3',
+                    profile: 'level1',
+                  },
+                ],
               },
             },
           ],
